refactor: seperate render function about cart
